Initialise SPECIAL so as to not end up in an error
SPECIAL is unitialised in `model_Init.f90` - seems like you can get
unlucky and start with it being `.TRUE.`, which will put you into a
'Negative integration time but not a special case' error and the program
will stop. SPECIAL should be `.FALSE.` until determined to be true by
reading the initial conditions.