Feature/enable rebuild (#1)
* Add gitignore

* Modify Makefile to start building with gfortran

* Rewrite some non-standard code

* Fix fixed format continuation issue

* Fix unary operator warnings and continuation problem

* Remove deprecated PAUSE statements

* Fix various incorrectnesses

- Ignore a few more file types that get created when making
- Fix a unary operator warning
- Correctly read input files to the end of the file
- Comment out a problematic, unused TUV subroutine

* Initialise SPECIAL so as to not end up in an error

SPECIAL is unitialised in `model_Init.f90` - seems like you can get
unlucky and start with it being `.TRUE.`, which will put you into a
'Negative integration time but not a special case' error and the program
will stop. SPECIAL should be `.FALSE.` until determined to be true by
reading the initial conditions.