Move unnecessary TUV-related files to temp directory
There were a bunch of files in the `tuv_old` directory that don't need
to be there. For the purposes of making the Docker image as minimal as
possible, these are going in a temporary `archive` directory until it
can be completely confirmed that they are unneeded to build and run the
model.